tools/clip-local-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
torch>=2.6.0
transformers==4.49.0
pillow==11.1.0
